Add styles to StockDetails component
